Move extension parser to utils
